split main.tsx into two parts to make react-refresh stop complaining about refresh fails when saving in main.tsx
