+++ src/main.tsx
@@ -1,31 +1,12 @@
-// Deps
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 // Locals
-import "./index.css";
-import { BearcaveRoot, AppRenderer } from "@components";
-import { Home } from "./homepage";
 
-function Bearcave(): JSX.Element {
-    return (
-        <Router>
-            <BearcaveRoot>
-                <Routes>
-                    <Route path="/*" element={<Home />} />
-                    <Route path="apps/*" element={<AppRenderer />} />
-                    <Route path="*" element={<PageNotFound />} />
-                </Routes>
-            </BearcaveRoot>
-        </Router>
-    );
-}
-
-function PageNotFound() {
-    return <div>Could not find the requested page.</div>;
-}
+import App from "./bearcave/app";
+import { createFirebase } from "./config";
 
 async function start(): Promise<void> {
+    // Ensure we have firebase up and running
+    createFirebase();
+
     /**
      * We dynamically import all index files under apps/* to include in bundle
      * //TODO: can this be done via vite config instead?
@@ -37,12 +18,7 @@ async function start(): Promise<void> {
 
     const container = document.getElementById("root");
     if (container) {
-        const root = createRoot(container);
-        root.render(
-            <StrictMode>
-                <Bearcave />
-            </StrictMode>,
-        );
+        App(container);
     } else {
         throw new Error("Bearcave failed to create root");
     }
